move nleigs types to their own module
